# Conflicts:
#	public/base.js
#	src/Controller/UpdateBBDDController.php
#	src/Controller/UploadProductController.php

+++ src/Controller/Remove.php
@@ -8,6 +8,7 @@ use Dflydev\FigCookies\SetCookie;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
+use SallePW\SlimApp\Model\Database\UserRepository;
 
 final class Remove
 {
@@ -24,7 +25,23 @@ final class Remove
 
     public function __invoke(Request $request, Response $response, array $args)
     {
+        /** @var UserRepository $repository */
         $repository = $this->container->get('user_repo');
+        if(!(isset($_SESSION['user_id']) && strlen($repository->findUserById($_SESSION['user_id']))) &&
+            !(isset($_COOKIE['user_id']) && strlen($repository->findUserById($_COOKIE['user_id'])))){
+            http_response_code(404);
+            die('Forbidden');
+        }
+
+        if(!isset($_SESSION['user_id'])){
+            $_SESSION['user_id'] = $_COOKIE['user_id'];
+        }
+
+        if($repository->isDeletedUser($_SESSION['user_id'])){
+            http_response_code(403);
+            die('Forbidden');
+        }
+
         $repository->deleteProduct($_POST['id_product']);
 
         $products = $this->container
@@ -45,7 +62,11 @@ final class Remove
 
             'products' => $products,
             'categ' => $categ,
-            'images' => $images
+            'images' => $images,
+            'profile_image' => $repository->getUserById($_SESSION['user_id'])->getProfileImage(),
+            'logged' => true,
+            'validated' => $repository->isValidated($_SESSION['user_id']),
+
 
         ]);
     }
